// src/rup/interfaces/rup-data.interface.ts
export interface RupData {
  tahun_anggaran: number;
  kd_klpd: string;
  nama_klpd: string;
  jenis_klpd: string;
  kd_satker: number;
  kd_satker_str: string;
  nama_satker: string;
  kd_rup: number;
  nama_paket: string;
  pagu: number;
  kd_metode_pengadaan: number;
  metode_pengadaan: string;
  kd_jenis_pengadaan: string;
  jenis_pengadaan: string;
  status_pradipa: string;
  status_pdn: string;
  status_ukm: string;
  alasan_non_ukm: string | null;
  status_konsolidasi: string;
  tipe_paket: string;
  kd_rup_swakelola: number | null;
  kd_rup_lokal: number | null;
  volume_pekerjaan: string;
  urarian_pekerjaan: string;
  spesifikasi_pekerjaan: string;
  tgl_awal_pemilihan: string;
  tgl_akhir_pemilihan: string;
  tgl_awal_kontrak: string;
  tgl_akhir_kontrak: string;
  tgl_awal_pemanfaatan: string;
  tgl_akhir_pemanfaatan: string;
  tgl_buat_paket: string;
  tgl_pengumuman_paket: string;
  nip_ppk: string;
  nama_ppk: string;
  username_ppk: string;
  status_aktif_rup: boolean;
  status_delete_rup: boolean;
  status_umumkan_rup: string;
  status_dikecualikan: boolean;
  alasan_dikecualikan: string | null;
  tahun_pertama: number | null;
  kode_rup_tahun_pertama: number | null;
  nomor_kontrak: string | null;
  spp_aspek_ekonomi: boolean;
  spp_aspek_sosial: boolean;
  spp_aspek_lingkungan: boolean;
  _event_date: string;
}